+++ src-tauri/src/models.rs
@@ -7,7 +7,7 @@ pub struct User {
     pub id: Option<i32>,
     pub email: String,
     pub display_name: String,
-    pub created_at: i32,
+    pub created_at: i64,
 }
 
 #[derive(Queryable, Selectable, QueryableByName)]
@@ -17,6 +17,6 @@ pub struct Note {
     pub id: Option<i32>,
     pub user_id: i32,
     pub content: Option<String>,
-    pub created_at: i32,
-    pub updated_at: i32,
+    pub created_at: i64,
+    pub updated_at: i64,
 }
